// petforest-fe-cms/src/utils/enums/input.enum.ts
export enum InputEnum {
  TEXT = 'text',
  NUMBER = 'number',
  THOUSAND_NUMBER = 'thousand-number',
  IMAGE = 'image',
  FILE = 'file',
  SELECT = 'select',
  SELECT_SEARCH = 'select-search',
  SELECT_INPUT = 'select-input',
  SELECT_AND_INPUT = 'select-and-input',
  SELECT_MULTI = 'select-multi',
  DATE_PICKER = 'date-picker',
  RANGE_PICKER = 'range-picker',
  PHONE = 'phone',
  HIDE = 'hide',
  TIME_PICKER = 'time-picker',
  SELECT_MEMBER = 'select-member',
  TEXT_AREA = 'text-area',
  OTHER = 'other',
  QUILL = 'quill',
}